requirements.txt
python-decouple
algoliasearch-django>=3.0,<4.0
django>=4.2.0,<5.0.0
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers